noticias/requirements.txt
attrs==20.3.0 
Automat==20.2.0 
cffi==1.14.4 
constantly==15.1.0 
cryptography==3.3.1  
cssselect==1.1.0  
cycler==0.10.0 
hyperlink==21.0.0 
idna== 3.1    
incremental==17.5.0 
itemadapter==0.2.0  
itemloaders==1.0.4  
jmespath==0.10.0 
kiwisolver==1.3.1  
lxml==4.6.2  
matplotlib==3.3.4  
mysql-connector-python==8.0.23 
numpy==1.20.0 
pandas==1.2.1  
parsel==1.6.0  
Pillow==8.1.0  
pip==20.0.2 
pkg-resources==0.0.0  
Protego==0.1.16 
protobuf==3.14.0 
pyasn1==0.4.8  
pyasn1-modules==0.2.8  
pycparser==2.20   
PyDispatcher==2.0.5  
PyHamcrest==2.0.2  
pyOpenSSL==20.0.1 
pyparsing==2.4.7  
python-dateutil==2.8.1  
pytz==2021.1 
queuelib==1.5.0  
Scrapy==2.4.1  
service-identity==18.1.0 
setuptools==44.0.0 
six==1.15.0 
Twisted==20.3.0 
w3lib==1.22.0 
zope.interface==5.2.0  